fixed length type in base64_decode and associated warning

+++ utils.cc
@@ -412,7 +412,7 @@ int base64_encode(const char *str, int length, char *b64store)
 }
 
 int
-base64_decode (const char *base64, int length, char *to)
+base64_decode(const char *base64, size_t length, char *to)
 {
   /* Table of base64 values for first 128 characters.  Note that this
      assumes ASCII (but so does Wget in other places).  */
@@ -426,7 +426,7 @@ base64_decode (const char *base64, int length, char *to)
     0,  0,  0, 63,  0, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40,
     41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51 
   };
-  int i;
+  size_t i;
   unsigned char* p = (unsigned char*) base64;
   char *q = to;
   int pad = length > 0 && (length % 4 || p[length - 1] == '=');
