+++ utils.h
@@ -237,7 +237,7 @@ char *b64enc(const unsigned char *data, int len);
 #define BASE64_LENGTH(len) (4 * (((len) + 2) / 3))
 int base64_encode(const char *str, int length, char *b64store);
 
-int base64_decode (const char *base64, int length, char *to);
+int base64_decode (const char *base64, size_t length, char *to);
 /* mmap() an entire file into the address space.  Returns a pointer
    to the beginning of the file.  The mmap'ed length is returned
    inside the length parameter.  If there is a problem, NULL is
